+++ admin-system/package.json
@@ -10,9 +10,11 @@
     "jquery": "^3.4.1",
     "object-hash": "^2.0.0",
     "react": "^16.10.1",
+    "react-beautiful-dnd": "^11.0.5",
     "react-bootstrap": "^1.0.0-beta.14",
     "react-csv-reader": "^1.3.1",
     "react-dom": "^16.10.1",
+    "react-file-reader": "^1.1.4",
     "react-icons": "^3.7.0",
     "react-loading": "^2.0.3",
     "react-router-dom": "^5.1.2",
